requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
groq>=0.9.0
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4
python-multipart==0.0.9
httpx==0.27.2
pydantic==2.8.2
lxml==5.3.0
openpyxl==3.1.5
python-dotenv==1.0.0
rank-bm25
